Employee Panel: What's pending: +Dialog fragment that would allow the user to choose their health. + Feed option on the employee screen

+++ app/src/main/java/com/s1mar/covid19tracker/utils/LoadingAnimationHelper.java
@@ -45,7 +45,7 @@ public class LoadingAnimationHelper {
 
     }
 
-    public static void dismissWithDelay(Activity activity, int millisecondsDelay){
+    public static void dismissWithDelay(Activity activity, long millisecondsDelay){
 
         new Handler().postAtTime(()->{
             ActLoaderBinding binder = initiateSplinterSequence(activity);
